requirements.txt
Django
django_recaptcha
numpy
Pillow
django-simpleui